requirements.txt
flask
Groq
python-dotenv
transformers
Flask-Dance
Flask-Login

